implement ServiceMeter and handle with TODO

+++ FileCabinetApp/Handlers/PurgeCommandHandler.cs
@@ -4,6 +4,8 @@ namespace FileCabinetApp.Handlers
 {
     public class PurgeCommandHandler : ServiceCommandHandlerBase
     {
+        private const RequestCommand Command = RequestCommand.Purge;
+        
         public PurgeCommandHandler(IFileCabinetService service) : base(service)
         {
         }
@@ -14,6 +16,17 @@ namespace FileCabinetApp.Handlers
         /// <param name="request"></param>
         public override void Handle(AppCommandRequest request)
         {
+            if (request is null)
+            {
+                throw new ArgumentNullException(nameof(request));
+            }
+            
+            if (request.Command != Command)
+            {
+                NextHandler.Handle(request);
+                return;
+            }
+            
             try
             { 
                 Service.Purge();
